[ADD] fix the parentuuid it still doens't work

+++ script/reset-database.sh
@@ -3,7 +3,9 @@
 set -e
 
 DB_CONTAINER=$(docker ps | grep 0.0.0.0:5432 | awk '{print $1}')
-docker stop "$DB_CONTAINER"
-docker rm "$DB_CONTAINER"
+if [ -n "$DB_CONTAINER" ]; then
+  docker stop "$DB_CONTAINER"
+  docker rm "$DB_CONTAINER"
+fi
 sudo rm -rf ~/.daf/data
 bash ./start-database.sh
